Moved PgHostRepository to src/repositories/pg/
